+++ app/src/main/java/com/example/vknewsclient/ui/theme/HomeScreen.kt
@@ -1,0 +1,82 @@
+package com.example.vknewsclient.ui.theme
+
+import androidx.compose.animation.core.tween
+import androidx.compose.foundation.ExperimentalFoundationApi
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.PaddingValues
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.lazy.LazyColumn
+import androidx.compose.foundation.lazy.items
+import androidx.compose.material3.ExperimentalMaterial3Api
+import androidx.compose.material3.SwipeToDismissBox
+import androidx.compose.material3.SwipeToDismissBoxValue
+import androidx.compose.material3.rememberSwipeToDismissBoxState
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.livedata.observeAsState
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.platform.LocalConfiguration
+import androidx.compose.ui.platform.LocalDensity
+import androidx.compose.ui.unit.dp
+import com.example.vknewsclient.MainViewModel
+
+@OptIn(ExperimentalMaterial3Api::class, ExperimentalFoundationApi::class)
+@Composable
+fun HomeScreen(
+    viewModel: MainViewModel,
+    paddingValues: PaddingValues,
+) {
+    val feedPosts = viewModel.feedPosts.observeAsState(listOf())
+    LazyColumn(
+        modifier = Modifier.padding(paddingValues),
+        contentPadding = PaddingValues(
+            top = 16.dp,
+            start = 8.dp,
+            end = 8.dp,
+            bottom = 72.dp
+        ),
+        verticalArrangement = Arrangement.spacedBy(8.dp)
+    ) {
+        items(feedPosts.value, key = {it.id}){feedPost->
+            val dismissThresholds =  with(LocalDensity.current){
+                LocalConfiguration.current.screenWidthDp.dp.toPx()*0.5F
+            }
+            val dismissBoxState = rememberSwipeToDismissBoxState(
+                positionalThreshold = {dismissThresholds},
+                confirmValueChange = {value ->
+                    val isDismissed = value in setOf(
+                        SwipeToDismissBoxValue.StartToEnd,
+                        SwipeToDismissBoxValue.EndToStart
+                    )
+                    if(isDismissed) viewModel.remove(feedPost)
+                    return@rememberSwipeToDismissBoxState isDismissed
+                }
+            )
+            SwipeToDismissBox(
+                modifier = Modifier.animateItemPlacement(animationSpec = tween(durationMillis = 300)),
+                enableDismissFromEndToStart = true,
+                enableDismissFromStartToEnd = false,
+                state = dismissBoxState,
+                backgroundContent = {},
+
+                ) {
+                PostCard(
+                    modifier = Modifier,
+                    feedPost = feedPost,
+                    onViewsClickListener = { statisticItem ->
+                        viewModel.updateCount(feedPost, statisticItem)
+                    },
+                    onLikesClickListener = { statisticItem ->
+                        viewModel.updateCount(feedPost, statisticItem)
+                    },
+                    onSharesClickListener = { statisticItem ->
+                        viewModel.updateCount(feedPost, statisticItem)
+                    },
+                    onCommentClickListener = { statisticItem ->
+                        viewModel.updateCount(feedPost, statisticItem)
+                    }
+                )
+            }
+
+        }
+    }
+}
