+++ app/src/main/java/com/example/vknewsclient/ui/theme/NavigationItem.kt
@@ -8,7 +8,6 @@ import androidx.compose.ui.graphics.vector.ImageVector
 import com.example.vknewsclient.R
 
 sealed class NavigationItem(val title: Int,val icon: ImageVector) {
-    data object Main:NavigationItem(title = R.string.home_title, icon = Icons.Outlined.Home)
     data object Favorite:NavigationItem(title = R.string.favorite_title, icon = Icons.Outlined.Favorite)
     data object Profile:NavigationItem(title = R.string.profile_title, icon = Icons.Outlined.Person)
 }
